seo + element by Houssem (#285)
* adding instance calculator

* Guarantee (#281)

* Update resource_management.json

Guaranty appears to be used as a verb here. Because of this it should be 'guarantee' https://www.grammar.com/guarantee_vs._guaranty

* Update disaster_recovery.json

When used as a verb it is recommended to use guarantee over guaranty

* adding boxboat tool

* fix typo

* adding link

* new test

* test3

* test 4

* Fix text

* adding CSS and target blank

* adding blank link

* lint css

* lint again

* important

* bloc pod access to IMDS within VMSS (#284)

* adding link to boxboat

* adding instance calculator

* Guarantee (#281)

* Update resource_management.json

Guaranty appears to be used as a verb here. Because of this it should be 'guarantee' https://www.grammar.com/guarantee_vs._guaranty

* Update disaster_recovery.json

When used as a verb it is recommended to use guarantee over guaranty

* adding boxboat tool

* fix typo

* adding link

* new test

* test3

* test 4

* Fix text

* adding CSS and target blank

* adding blank link

* lint css

* lint again

* important

Co-authored-by: msftph <65667413+msftph@users.noreply.github.com>

* Bloc Pod access to VMSS IMDS

I didn't find the good documentation that refers to how to fix this and how to create the Network Policy for this case.

Co-authored-by: Louis-Guillaume MORAND <lgm@lgmorand.com>
Co-authored-by: msftph <65667413+msftph@users.noreply.github.com>

* optimize SEO

Co-authored-by: msftph <65667413+msftph@users.noreply.github.com>
Co-authored-by: Houssem Dellai <houssem.dellai@live.com>